Move log flusher to separate file
